refact: moves reverse shifts to decrypter and updates crypt removing it
