+++ lib/decrypter.rb
@@ -13,4 +13,8 @@ class Decrypter < Crypt
     { decryption: restore_spec_chars(decrypted_message, crypt_data[:spec_chars]),
       key: crypt_data[:converted_key].key, date: format_date(date).to_s }
   end
+
+  def reverse_shifts(shifts)
+    shifts.map { |shift| -shift }
+  end
 end
