+++ spec/spec_helper.rb
@@ -8,5 +8,6 @@ require_relative 'enigma_spec'
 require_relative 'key_spec'
 require_relative 'crypt_spec'
 require_relative 'encrypter_spec'
+require_relative 'decrypter_spec'
 require_relative 'timeable_spec'
 require_relative 'stringable_spec'
